[ 仕様変更 ] スライダーのページネーションに ltg-slide-pagination を追加し移行先案内をテーマ所有クラスに変更
changelog の移行先を Swiper 所有クラスからテーマ管理下の ltg-slide-* に差し替え、
エントリを 2 行以内に短縮。廃止バージョンのコメントを実測できる事実に修正。
テストのフィクスチャを trait に切り出して preload テストとの重複を解消し、
接頭辞のエスケープを検知する境界値ケースを追加。
